Extract sandbox HTTP stream error handling
